Updated Kenny's Demo Flask
Updated Kenny's Demo Flask to use the new HTML pages added.  

If you do plan on testing this on your own, you'll need to put both of Bobby's and Meghan HTML files under a templates folder so that the render_template function can call them, along with creating a /static/css folder and put Bobby's CSS file in it.  

If you need help setting it up, just shoot me a message on GroupMe