Scuffed login system
Currently hard coded the logins for specific users and redirected them to their certain pages.  If we want to go down this path, I'll need the users logins we want in the system.    Still attempting at working on using a database with it, will probably need help with that.

Also changed the loginpage button to a submit button